+++ crates/sdk/src/js/mod.rs
@@ -25,6 +25,9 @@ pub mod utils;
 /// Events.
 pub mod events;
 
+/// GLV.
+pub mod glv;
+
 /// Simulation.
 pub mod simulation;