+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(sdk): Added shift simulation.
 - sdk(sdk): Added GLV deposit simulation.
 - sdk(sdk): Added GLV withdrawal simulation.
+- sdk(sdk): Added JsGlv and JsGlvModel.
 - cli: Added auto-creation support of multiple ALTs based on ALT limit for `alt extend` command.
 
 ### Changed